Relocate all egui components to files
